+++ .github/workflows/cd-packaging-tests/bundler-webpack/webpack.config.js
@@ -6,6 +6,9 @@ module.exports = {
       stream: require.resolve("stream-browserify"),
       util: require.resolve("util/"),
       buffer: require.resolve("buffer/"),
+      os: require.resolve("os-browserify/browser"),
+      assert: require.resolve("assert/"),
+      zlib: require.resolve("browserify-zlib"),
     },
   },
 };